+++ test/ReplSpec.hs
@@ -25,12 +25,15 @@ commands =
     [ ReadInputRegistersWord
     , ReadInputRegistersBits
     , ReadInputRegistersFloat
+    , ReadInputRegistersDouble
     , ReadHoldingRegistersWord
     , ReadHoldingRegistersBits
     , ReadHoldingRegistersFloat
+    , ReadHoldingRegistersDouble
     , WriteRegistersWord
     , WriteRegistersBits
     , WriteRegistersFloat
+    , WriteRegistersDouble
     , Read
     , Write
     , StartHeartbeat