+++ ChangeLog.md
@@ -4,5 +4,6 @@
 
 ## 0.1.1
 
+- Add Double (4 words) type to ModValue
 - Display parse errors as notifications
 - Register table display tuning